ALMOOL_Chat/requirements.txt
Flask==2.3.3
requests==2.31.0
Jinja2==3.1.2
Werkzeug==2.3.7
python-dotenv==1.0.0
openai>=1.0.0
google-genai
google-generativeai
pycryptodome
